requirements.txt
appnope==0.1.2
asgiref==3.3.1
backcall==0.2.0
db-question @ file:///Users/orangewater/test_work/db_question/dist/db_question-0.1.tar.gz
decorator==4.4.2
Django==3.1.4
django-cors-headers==3.6.0
django-extensions==3.1.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
ipython==7.19.0
ipython-genutils==0.2.0
jedi==0.18.0
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.8
ptyprocess==0.7.0
Pygments==2.7.3
PyJWT==1.7.1
pytz==2020.5
sqlparse==0.4.1
traitlets==5.0.5
wcwidth==0.2.5
